Clean up Create Recipe components, remove unused parameters from form elements
