+++ client/source/components/CreateRecipe/AddIngredient.jsx
@@ -1,5 +1,13 @@
 import React from 'react';
 
+//Bootstrap 
+import { Grid, Row, Col, FormGroup, FormControl, Button, Jumbotron, Carousel, Container, ControlLabel } from 'react-bootstrap';
+
+var units = ['select unit', 'tsp.', 'tbsp.', 'fl.oz.', 'cup', 'pt.', 'qt.', 'gal.', 'g', 'kg', 'oz.', 'lbs', 'whole']; 
+var volume = ['tsp.', 'tbsp.', 'fl.oz.', 'cup', 'pt.', 'qt.', 'gal.']; 
+var mass = ['g', 'kg', 'oz.', 'lbs']; 
+var other = ['whole']; 
+
 class AddIngredient extends React.Component {
 	constructor(props) {
 		super(props);
@@ -16,6 +24,8 @@ class AddIngredient extends React.Component {
 	}
 
 	componentWillMount(){
+		console.log('Add Ingredient will mount!'); 
+		// console.log('ADD INGREDIENT POSITION:', this.props.number); 
 		this.setState({
 			position: this.props.number
 		}); 
@@ -44,34 +54,52 @@ class AddIngredient extends React.Component {
 
 	render () {
 		return (
-			<div>
-				<div className="row">
-					<div className="input-field col s4">
-					  <input placeholder={this.state.ingredients} id="name" type="text" className="validate" onChange={this.handleChange.bind(this)}/>
-					  <label htmlFor="servings">Ingredient</label>
-					</div>
-					<div className="input-field col s4">
-					  <input id="amount" type="text" className="validate" onChange={this.handleChange.bind(this)}/>
-					  <label htmlFor="servings">Amount</label>
-					</div>
-					<div className="input-field col s4">
-					  <input id="unit" type="text" className="validate" onChange={this.handleChange.bind(this)}/>
-					  <label htmlFor="servings">Unit</label>
-					</div>
-				</div>
-				<div className="row">
-					<div className="input-field col s6">
-					  <input placeholder='Preparation' id="prep" type="text" className="validate" onChange={this.handleChange.bind(this)}/>
-					  <label htmlFor="servings">Preparation</label>
-					</div>
-					<div className="input-field col s6">
-					  <input placeholder='Additional Notes' id="optional" type="text" className="validate" onChange={this.handleChange.bind(this)}/>
-					  <label htmlFor="servings">Notes</label>
-					</div>
-				</div>
-				<button onClick={this.handleClick.bind(this)}> Next Step </button> 
-			</div>
-		); 
+			<Grid>
+			  <form onSubmit={this.handleClick.bind(this)}>
+			<Row className="show-grid">
+			  <Col xs={4} md={4}> 
+			  		<FormGroup>
+			        <ControlLabel> IngredientName </ControlLabel>
+			        <FormControl type="text" id="name" onChange={this.handleChange.bind(this)} value={this.state.name} />
+			        </FormGroup>
+			  </Col>
+			  <Col xs={4} md={4}> 
+			  		<FormGroup>
+			        <ControlLabel> Quantity </ControlLabel>
+			        <FormControl type="text" id="amount" onChange={this.handleChange.bind(this)} value={this.state.amount} />
+			        </FormGroup>
+			  </Col>
+			  <Col xs={4} md={4}> 
+		          <FormGroup>
+		            <ControlLabel>Units</ControlLabel>
+		            <FormControl componentClass="select" id="unit" onChange={this.handleChange.bind(this)} >
+		              {units.map((unit, i)=> (
+		              	<option key={'unitDropdown' + i + this.state.position} value={unit}>{unit}</option>
+		              ))}
+		            </FormControl>
+		          </FormGroup>
+			  </Col>
+			</Row>
+			<Row className="show-grid" style={{padding: 5}}>
+			  <Col xs={6} md={6}> 
+			  		<FormGroup>
+			        <ControlLabel> Preparation </ControlLabel>
+			        <FormControl type="text" id="prep" onChange={this.handleChange.bind(this)} value={this.state.prep} />
+			        </FormGroup>
+			  </Col>
+			  <Col xs={6} md={6}> 
+			  		<FormGroup>
+			        <ControlLabel> Additional Notes </ControlLabel>
+			        <FormControl type="text" id="optional" onChange={this.handleChange.bind(this)} value={this.state.optional} />
+			        </FormGroup>
+			  </Col>
+			</Row>
+			<Row className="show-grid">
+			  <button style={{display: 'none'}} type="submit"onClick={this.handleClick.bind(this)}> Next Step </button> 
+			</Row>
+			</form>
+			</Grid>
+		); 	
 	}
 }
