+++ client/source/components/CreateRecipe/IngredientsForm.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AddIngredient from './AddIngredientBS'; 
+import AddIngredient from './AddIngredient'; 
 import RecipeIngredients from '../Recipe/RecipeIngredientsBS'; 
 
 class IngredientsForm extends React.Component {
@@ -42,8 +42,8 @@ class IngredientsForm extends React.Component {
 			<div>
 				<h3> Recipe Ingredients </h3>
 				<RecipeIngredients ingredientList={this.state.ingredients} />
-				{this.state.newIngredient.map((ingredient) => (
-					<AddIngredient key={'ingredient' + this.state.ingredientsCount} number={this.state.ingredientsCount} handleAddIngredient={this.handleAddIngredientForm.bind(this)} ingredient={ingredient}/>
+				{this.state.newIngredient.map((ingredient, i) => (
+					<AddIngredient key={'ingredient' + i} number={this.state.ingredientsCount} handleAddIngredient={this.handleAddIngredientForm.bind(this)} ingredient={ingredient}/>
 				))}
 			</div>
 		);
